# Adding plugin to lazy load YouTube videos on features page