--- a/docs/theme/static/styles/_lazy-youtube.less
+++ b/docs/theme/static/styles/_lazy-youtube.less
@@ -1,0 +1,58 @@
+.youtube-container {
+    display: block;
+    margin: 0 auto;
+    margin: 20px 0;
+    width: 100%;
+    max-width: 600px;
+    .youtube-player {
+        display: block;
+        width: 100%;
+        padding-bottom: 75%;
+        overflow: hidden;
+        position: relative;
+        width: 100%;
+        height: 100%;
+        cursor: hand;
+        cursor: pointer;
+        display: block;
+        .youtube-thumb {
+            bottom: 0;
+            display: block;
+            left: 0;
+            max-width: 100%;
+            width: 100%;
+            position: absolute;
+            right: 0;
+            top: 0;
+            height: auto
+        }
+        .play-button {
+            height: 48px;
+            width: 68px;
+            left: 50%;
+            top: 50%;
+            margin-left: -34px;
+            margin-top: -24px;
+            position: absolute;
+            svg:hover {
+                .yt-play-button-bg  {
+                    fill: #cc181e;
+                }
+            }
+        }
+        &.active .yt-play-button-bg  {
+            fill: #cc181e;
+        }
+        .youtube-iframe {
+            width: 100%;
+            height: 100%;
+            position: absolute;
+            top: 0;
+            left: 0;
+        }
+    }
+}
+
+body:not(.donejs):not(.community) > .container-fluid .content.docs .comment img.youtube-thumb {
+    margin: 0;
+}