--- a/docs/theme/static/styles/_shared_layout.less
+++ b/docs/theme/static/styles/_shared_layout.less
@@ -227,7 +227,7 @@ body:not(.donejs):not(.community) {
 					a {
 						color: @logoColor;
 					}
-					img {
+					.fun-img {
 						display: inline-block;
 						margin: -20px 0 0 0 !important;
 						max-height: 80px;