--- a/docs/theme/static/styles/styles.less
+++ b/docs/theme/static/styles/styles.less
@@ -21,6 +21,7 @@
 @import '_header_nav.less';
 @import '_misc_sitewide.less';
 @import '_animation.less';
+@import '_lazy-youtube.less';
 
 //files for the home page "body.donejs"
 @import '_donejs-typography.less';