This greatly reduces the # of HTTP requests that happen on page load. Here are the basic performance metrics:

Metric | Before | After
--------|--------|-------
__Requests__ | 109 | 72
__Finish__ | 2.51s | 746ms
__DOMContentLoaded__ | 489ms | 241ms
__Load__ | 2.41s | 734ms

Fixes #364